feat: add memory allocation to install and ingress expiry to agent (#30)
replica api:
- added `ingress_expiry: u64` to `CallRequest` `RequestStatusRequest` `QueryRequest`
- made `RequestStatusResponse` a struct with a `pub time: u64` field  and moved statuses to new `Status` enum with new `Done` status
- add `time: 64` as a part of the `RequestStatusResponse` struct, agent doesn't do anything with this time value right now

rust agent:
- add `ing_exp_duration: Duration` to `Agent` struct with default Duration of 5 minutes
- add private `get_expiry_date(&self) -> u64` to `Agent` to compute the ingress expiry date with `ing_exp_duration` value
- added `ing_exp_duration: Option<std::time::Duration>` to `AgentConfig`
- added `pub fn expire_after(self, duration: Option<std::time::Duration>)` to `AgentBuilder` this sets the `ing_exp_duration` field in `AgentConfig`

- added `ing_exp_datetime: Option<u64>` to `UpdateBuilder` this field is the "upper limit on the validity of the request, expressed in nanoseconds since 1970-01-01 (like ic0.time())." as noted in the Public Spec.
If the user creates an update request using `UpdateBuilder`, this value will be used instead of the ingress expiry date from `Agent::get_expiry_date()`
- added two methods to `UpdateBuilder`:
- `pub fn expire_at(&mut self, time: std::time::SystemTime)` 
Which takes a SystemTime. Converts it to a Duration by calling duration_since(UNIX_EPOCH) to learn about where in time this SystemTime lies. The Duration is converted to nanoseconds and stored in `ing_exp_datetime`
- `pub fn expire_after(&mut self, duration: std::time::Duration)`
Takes a Duration (i.e. 30 sec/5 min 30 sec/1 h 30 min, etc.) and adds it to the Duration of the current SystemTime since the UNIX_EPOCH. Converts the sum to nanoseconds and stores in `ing_exp_datetime`
- made `update_raw` private, removing it from the public api


management canister and canister attributes:
- add `MemoryAllocation` to install method and to canister attributes

spec compliance:
- compliance with `0.10.2`

misc.
- use `thiserror` for `ComputeAllocationError` and `MemoryAllocationError`

Closes #32