feat: ic-utils crate which provides utilities for communicating with the Internet Computer (#29)
The intent is to add higher level functions in this library, with more dependencies, while keeping the base `ic-agent` crate as simple as possible. `ic-utils` will also be used on the Internet Computer to allow for inter-canister calls.

Currently, it encapsulates Canister calls and allow creation of a canister interfaces. See the ManagementCanister to see how an interface can be implemented (including code generated from Candid files).