feat: add better with_XXX methods for builders (#49)
Added a with_boxed_identity method to AgentBuilder for when the caller does not know the type. Did the same for
password_manager since it's the same kind of API.

Also added a version of allocations that take an Option, instead
of having the callee use "let Some(ma) = ..." when all they have
is an option.