feat: add a ReplicaV1Transport trait to communicate with the replica (#122)
A transport is a specialized facade and it applies better here.

This also simplifies the requirements for values being sent/received to be `Send` but not `Sync`. They don't need to exist on multiple threads/fibers so we can relax that requirement. This makes the Transport a bit easier.

This also added the Clone trait for Agent. Cloning is cheap now as the transport is ref counted and doesn't need to be cloned.

# BREAKING CHANGE

This whole PR is a breaking change if you're using the previous API. The best you can do is use the `reqwest` feature and ignore the deprecation warning for `with_url`.