fix: ic-asset::sync() accepts multiple source directories (#355)
* This is a breaking change

fix: ic-asset::sync() accepts multiple source directories

It also now skips filenames and directories that begin with "." (as dfx does),
and reports an error if more than one asset resolves to the same key.

This is to account for asset source directories that may in the future contain
a configuration file .ic-asset.json.  More than one source directory may contain
such a file, but its configuration should only apply to the assets in the
corresponding directory.